Rewrite of TileLayer: Solve problems + perf gain!
 * `Pipeline` is renamed to `CommandBuffer`
 * Moved tile related stuff into its own folder
 * New rendering of tiles:
   * Draw only what's visible
   * No seams!
   * No pixels errors!
   * Super fast!!
   * =D
 * NOTE: at this point camera is totally broken. :'(
