+++ editor/gl/gl.ts
@@ -3,3 +3,4 @@ export * from './program';
 export * from './buffer';
 export * from './tex';
 export * from './draw';
+export * from './quad';
