+++ editor/rendering/pipeline.ts
@@ -1,4 +1,4 @@
-import {PipelineEl, TextureGetter} from './interfaces';
+import {Command, TextureGetter} from './interfaces';
 import {Context} from './context';
 
 /// The clear all pipeline element performs
@@ -16,7 +16,7 @@ export const FlipY = (ctx: Context) => {
 };
 
 /// Skip the node if the condition function returns true
-export const SkipIf = (el: PipelineEl, test: (ctx: Context) => boolean) => {
+export const SkipIf = (el: Command, test: (ctx: Context) => boolean) => {
     return (ctx: Context) => {
         if (!test(ctx)) {
             el.execute(ctx);
@@ -26,9 +26,9 @@ export const SkipIf = (el: PipelineEl, test: (ctx: Context) => boolean) => {
 
 /// Change the filtering temporarily
 /// of the passed object.
-export class TmpLinearFiltering implements PipelineEl {
+export class TmpLinearFiltering implements Command {
     constructor(
-        private render_el: TextureGetter & PipelineEl
+        private render_el: TextureGetter & Command
     ) {}
 
     execute(ctx: Context) {
@@ -63,27 +63,27 @@ export class TmpLinearFiltering implements PipelineEl {
     }
 }
 
-type PipelineElLambda = (ctx: Context) => void;
+type CommandLambda = (ctx: Context) => void;
 
 /// The pipeline class, this is where the execution
 /// takes place. This is also the context owner.
-export class Pipeline {
+export class CommandBuffer {
 
     private context: Context = undefined;
-    private raw: PipelineEl[];
+    private raw: Command[];
 
     constructor(
-        raw: Array<PipelineEl | PipelineElLambda>
+        raw: Array<Command | CommandLambda>
     ) {
         this.raw = raw.map(el => {
             if (typeof el == 'function') {
                 // TODO: Remove 'as' in next version of TypeScript
                 return {
-                    execute: el as PipelineElLambda,
+                    execute: el as CommandLambda,
                 };
             } else {
                 // TODO: Remove 'as' in next version of TypeScript
-                return el as PipelineEl;
+                return el as Command;
             }
         });
     }
