+++ editor/rendering/tiles/index.ts
@@ -1,0 +1,4 @@
+
+export {TilesLayerBuilder, TilesHandle, SelectedPartialLayer} from './common';
+export {TilesLayer} from './vbo';
+export {TilesLayer2} from './quad';
