+++ package.json
@@ -38,7 +38,7 @@
     "ts-loader": "0.8.1",
     "tslint": "3.6.0",
     "tslint-loader": "2.1.3",
-    "typescript": "1.8.9",
+    "typescript": "1.8.10",
     "typings": "0.7.9",
     "url-loader": "0.5.6",
     "webpack": "1.12.14",
