+++ tsconfig.json
@@ -56,6 +56,7 @@
         "./editor/gl/draw.ts",
         "./editor/gl/gl.ts",
         "./editor/gl/program.ts",
+        "./editor/gl/quad.ts",
         "./editor/gl/tex.ts",
         "./editor/gl/uniform.ts",
         "./editor/index.ts",
@@ -72,6 +73,11 @@
         "./editor/rendering/shaders.ts",
         "./editor/rendering/sprite.ts",
         "./editor/rendering/tiles.ts",
+        "./editor/rendering/tiles/common.ts",
+        "./editor/rendering/tiles/index.ts",
+        "./editor/rendering/tiles/layer.ts",
+        "./editor/rendering/tiles/quad.ts",
+        "./editor/rendering/tiles/vbo.ts",
         "./editor/rendering/util.ts",
         "./editor/rendering/viewport.ts",
         "./editor/rules/parser.ts",
